% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_numCs_numTs_region.R
\name{aggregate_numCs_numTs_region}
\alias{aggregate_numCs_numTs_region}
\title{Computes the aggregate methylation for a specified genomic region across 
samples, grouped by their respective conditions.}
\usage{
aggregate_numCs_numTs_region(region_res, region, conditions, med = FALSE)
}
\arguments{
\item{region_res}{A nested list where each top-level element is a sample 
(e.g., `region_res$SRR28431752$18S$numCs`).}

\item{region}{A string specifying the region of interest (e.g., `"18S"`).}

\item{conditions}{A named list where each name is a condition 
(e.g., `OLD`, `YOUNG`) and each element is a vector of sample IDs belonging 
to that condition.}

\item{med}{A logical value indicating whether to compute the median 
(if `TRUE`) or the mean (if `FALSE`) for the aggregate methylation values. 
Default is `FALSE`.}
}
\value{
A dataframe with columns:
\describe{
  \item{sample_ID:}{The sample identifier.}
  \item{condition:}{The condition to which the sample belongs.}
  \item{aggregate_meth:}{The aggregate methylation value for the 
  specified region.}
}
}
\description{
Computes the aggregate methylation for a specified genomic region across 
samples, grouped by their respective conditions.
}
