% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rDNA_data.R
\name{get_rDNA_data}
\alias{get_rDNA_data}
\title{Extract rDNA Chromosome Data}
\usage{
get_rDNA_data(meth_object, num_samples, rDNA_chr_name = "rDNAm")
}
\arguments{
\item{meth_object}{A methylation object (e.g., from `methylKit`).}

\item{num_samples}{Number of samples in `meth_object` to process.}

\item{rDNA_chr_name}{Name of the rDNA chromosome (default: `"rDNAm"`).}
}
\value{
A subset of `meth_object` containing only the rDNA chromosome data.
}
\description{
Extracts all data associated with the rDNA chromosome from a `meth_object`.
}
\examples{
# Load the example mouse WGBS dataset
file_path <- system.file(
  "extdata",
  "filtered_meth_raw_chr1_chrRDNAm.rds",
  package = "myoRDNA",
  mustWork = TRUE
)

loaded_data <- readRDS(file_path)

# Define experimental conditions
# 0 = OLD, 1 = YOUNG
treatment <- c(
  0, 0, 0, 0, 0, 0,
  1, 1, 1, 1
)

num_samples <- length(treatment)

# Construct a methylRawList object
loaded_data <- methylKit::methylRawList(
  loaded_data,
  treatment = treatment
)

# Extract methylation data for the rDNA chromosome
meth_rDNA <- get_rDNA_data(
  meth_object = loaded_data,
  num_samples = num_samples,
  rDNA_chr_name = "chrRDNAm"
)

# Inspect the extracted data
print(meth_rDNA)

}
