% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rDNA_data.R
\name{get_rDNA_data}
\alias{get_rDNA_data}
\title{Extract rDNA Chromosome Data}
\usage{
get_rDNA_data(meth_object, num_samples, rDNA_chr_name = "rDNAm")
}
\arguments{
\item{meth_object}{A methylation object (e.g., from `methylKit`).}

\item{num_samples}{Number of samples in `meth_object` to process.}

\item{rDNA_chr_name}{Name of the rDNA chromosome (default: `"rDNAm"`).}
}
\value{
A subset of `meth_object` containing only the rDNA chromosome data.
}
\description{
Extracts all data associated with the rDNA chromosome from a `meth_object`.
}
\examples{
get_rDNA_data(meth_object, num_samples = 10, rDNA_chr_name = "rDNAm")
}
